detail and profile have passed, form is almost done
