+++ pages/[id]/form/index.jsx
@@ -1,26 +1,30 @@
 import React, { useEffect, useContext } from 'react';
-import { TokenContext } from '../utils/context';
-import LoadingDots from '../components/loading';
-import Button from '../components/button';
-import Layout from '../components/Layout';
-import Input from '../components/input';
+import { TokenContext } from '../../../utils/context';
+import LoadingDots from '../../../components/loading';
+import Button from '../../../components/button';
+import Layout from '../../../components/Layout';
+import Input from '../../../components/input';
 import { useState } from 'react';
+import { useRouter } from "next/router";
 import Image from 'next/image';
+import Link from 'next/link';
 
-import Photo from '../assets/foto-form.png';
-import BRI from '../assets/bri.png';
-import BCA from '../assets/bca.png';
+// import Photo from '../assets/foto-form.png';
+import BRI from '../../../assets/bri.png';
+import BCA from '../../../assets/bca.png';
 
-const Form = (id) => {
+const Form = (props) => {
 
   const [loading, setLoading] = useState(true);
   const { token } = useContext(TokenContext);
 
   const [data, setData] = useState([]);
   const [account, setAccount] = useState({});
-  const [date, setDate] = useState({});
+  const [startDate, setStartDate] = useState('10-12-2022');
+  const [endDate, setEndDate] = useState('12-12-2022');
   const [item, setItem] = useState({});
-  const [bank, setBank] = useState('');
+  const [bank, setBank] = useState('BCA');
+  const router = useRouter();
 
   const handleChange = (e) => {
     const name = e.target.name;
@@ -33,40 +37,40 @@ const Form = (id) => {
     console.log(account);
   };
 
+  useEffect(() => {
+    fetchData();
+  },[])
+
   setBank = (e) => {
     console.log(e.target.value);
   };
-
-  useEffect(() => {
-    fetchData();
-  }, [])
 
   const fetchData = () => {
     let myHeaders = new Headers();
     myHeaders.append(`Authorization`, `Bearer ${token}`);
     myHeaders.append(`Content-Type`, `application/json`);
 
-    let raw = JSON.stringify({
-      "rooms_id": 10,
-      "date_start": "16-09-2022",
-      "date_end": "18-09-2022",
-      "bank": "BCA"
-    });
-
+    console.log("ini data di form", router);
+    if (router.query.id == undefined) {
+      // return fetchData();
+      return false
+    }
+  
     let requestOptions = {
       method: 'GET',
       headers: myHeaders,
-      body: raw,
       redirect: 'follow'
     };
 
-    fetch(`https://mnroom.capstone.my.id/rents/${id}`, requestOptions)
+
+    fetch(`https://mnroom.capstone.my.id/rents/${router.query.id}`, requestOptions)
       .then(response => response.json())
       .then((result) => {
         setData(result.data);
+        console.log("ini data get di form", result.data);
       })
       .catch(error => console.log('error', error))
-      .finally(() => handleConfirm());
+      .finally(() => setLoading(false));
   }
 
   const handleConfirm = () => {
@@ -75,10 +79,10 @@ const Form = (id) => {
     myHeaders.append(`Content-Type`, `application/json`);
 
     let raw = JSON.stringify({
-      "rooms_id": 6,
-      "date_start": "2022-11-11",
-      "date_end": "2022-11-12",
-      "bank": "BNI"
+      "rooms_id": parseInt(router.query.id),
+      "date_start": startDate,
+      "date_end": endDate,
+      "bank": bank
     });
 
     let requestOptions = {
@@ -92,14 +96,20 @@ const Form = (id) => {
       .then(response => response.json())
       .then((result) => {
         alert(result.message);
+        console.log('ini data result'. result.message);
       })
       .catch(error => console.log('error', error))
       .finally(() => setLoading(false));
   }
 
-  const handleDate = (e) => {
+  const handleStartDate = (e) => {
     const inputDate = e.target.value;
-    setDate({ ...date, [e.target.name]: inputDate });
+    setStartDate({ ...date, [e.target.name]: inputDate });
+  }
+
+  const handleEndDate = (e) => {
+    const inputDate = e.target.value;
+    setEndDate({ ...date, [e.target.name]: inputDate });
   }
 
   const handleBank = (e) => {
@@ -136,7 +146,7 @@ const Form = (id) => {
             <Input
               shrink={true}
               className="text-black form w-full input pb-2"
-              onChange={(e) => handleDate(e)}
+              onChange={(e) => handleStartDate(e)}
               label="Start Date"
               type='date'
               name="stardate"
@@ -148,7 +158,7 @@ const Form = (id) => {
             <Input
               shrink={true}
               className="text-black form w-full input pb-2"
-              onChange={(e) => handleDate(e)}
+              onChange={(e) => handleEndDate(e)}
               label="End Date"
               type='date'
               name="enddate"
@@ -181,11 +191,13 @@ const Form = (id) => {
             <br />
 
             <div className='text-end pb-10'>
+               <Link href={`/${router.query.id}/payment`}>
               <Button
                 className=" bg-orange-600 hover:bg-orange-400 font-bold py-2 px-2 mb-3 rounded text-white"
                 label="Confirm"
                 onClick={() => handleConfirm()}
               />
+              </Link>
             </div>
 
           </div>
